extract link.ParseCollection
